moved all uianimator2 options to uianimator2_options.py
